Moved implementation to new file.
